Added new feature for high score.
Created a new text file called 'high_score.txt'. Stored a default value
of 0 (zero) in the file.

Modified declarations.py file to remove redundant lines.

Modified main.py file so that the high score is read from the text file
and if the current game score is higher than the score in the text file,
then the new higher score will be written in the text file.

Modified home_page() function in methods.py file so that high score is
displayed on the game screen.